saluki_metrics

Trait MetricTag

pub trait MetricTag: Sealed {
    // Required method
    fn into_label(self) -> Label;
}

A metric tag.

Marker trait for types which can support being used as a metric tag, in order to optimize their internal representation to avoid unnecessary allocations.

This trait is sealed and cannot be implemented outside of this crate.

Required Methods

fn into_label(self) -> Label

Consumes self and converts it to a tag.

Under the hood, the metrics crate is used, which calls tags “labels” instead, which is where the return type naming comes from.